Removing TypeScript template
This PR simply removes the TypeScript template, since it's currently outdated (and it's not even using Express), and until we get more user feedback, I'd like to lead with just a single Express template in order to reduce confusion around which template to select for Node.js. When looking at the telemetry, this template doesn't get used very much, so it also shouldn't be impactful when removed.